+++ mobile/src/app/navigation/stacks/ProfileStack.tsx
@@ -1,13 +1,10 @@
 import React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import {Profile} from 'pages';
 import {Screens} from '../navigationEnums';
 import {ProfileStackParamsList} from '../navigationTypes';
-import {PaymentMethods} from 'pages/PaymentMethods';
-import {Settings} from 'pages/Settings';
-import {Orders} from 'pages/Orders';
+import {PaymentMethods, Orders, Settings, Profile} from 'pages';
 
-const ProfileStackNavigator = () => {
+export const ProfileStackNavigator = () => {
   const ProfileStack = createNativeStackNavigator<ProfileStackParamsList>();
 
   return (
@@ -35,5 +32,3 @@ const ProfileStackNavigator = () => {
     </ProfileStack.Navigator>
   );
 };
-
-export default ProfileStackNavigator;
