+++ mobile/src/app/navigation/stacks/index.ts
@@ -1,13 +1,5 @@
-import MainTabsNavigator from './MainTabsStack';
-import HomeStackNavigator from './HomeStack';
-import CatalogStackNavigator from './CatalogStack';
-import MapStackNavigator from './MapStack';
-import ProfileStackNavigator from './ProfileStack';
-
-export {
-  MainTabsNavigator,
-  HomeStackNavigator,
-  CatalogStackNavigator,
-  MapStackNavigator,
-  ProfileStackNavigator,
-};
+export * from './MainTabsStack';
+export * from './HomeStack';
+export * from './CatalogStack';
+export * from './MapStack';
+export * from './ProfileStack';
