+++ mobile/src/entities/filter/model/types.ts
@@ -1,0 +1,8 @@
+import {ImageSourcePropType} from 'react-native';
+
+export type Filter = {
+  id: string;
+  title: string;
+  image: ImageSourcePropType;
+  keyWord: string;
+};
