+++ mobile/src/entities/product/index.ts
@@ -1,0 +1,1 @@
+export * from './model/types';
