+++ mobile/src/entities/product/model/types.ts
@@ -1,0 +1,12 @@
+import {ImageSourcePropType} from 'react-native';
+
+export type Product = {
+  id: string;
+  title: string;
+  brand: string;
+  price: number;
+  totalScore: number;
+  reviewsCount: number;
+  priceWithDiscount: number;
+  image: ImageSourcePropType;
+};
