--- mobile/src/pages/PaymentMethods/index.ts
@@ -1,1 +1,0 @@
-export * from './ui';
