--- mobile/src/pages/Settings/index.ts
@@ -1,1 +1,0 @@
-export * from './ui';
