+++ mobile/src/pages/index.ts
@@ -4,3 +4,6 @@ export * from './Map/ui';
 export * from './Profile/ui';
 export * from './Cart/ui';
 export * from './Auth/ui';
+export * from './Settings/ui';
+export * from './Orders/ui';
+export * from './PaymentMethods/ui';
