+++ mobile/src/shared/assets/mockData/index.ts
@@ -127,8 +127,10 @@ export const products = [
     title: 'Поло',
     image: require('../images/mock/tshirt.png'),
     price: 10000,
-    priceWithDiscount: '6999',
+    priceWithDiscount: 6999,
     brand: 'Ralpf Loren',
+    reviewsCount: 12,
+    totalScore: 3,
   },
   {
     id: '2',
@@ -136,18 +138,20 @@ export const products = [
     image: require('../images/mock/tshirt.png'),
     price: 25000,
     brand: 'Husqvarna AB',
+    reviewsCount: 1442,
+    totalScore: 4.9,
   },
   {
     id: '3',
     title: 'iPhone 15 pro',
     image: require('../images/mock/tshirt.png'),
     price: 122000,
-    priceWithDiscount: '$10',
+    priceWithDiscount: 99999,
     brand: 'Apple',
   },
   {
     id: '4',
-    title: 'Я без бренда',
+    title: 'Без бренда и отзывов',
     image: require('../images/mock/tshirt.png'),
     price: 4999,
   },
@@ -156,7 +160,9 @@ export const products = [
     title: 'Длинное наименование товара',
     image: require('../images/mock/tshirt.png'),
     price: 500,
-    priceWithDiscount: '$10',
+    priceWithDiscount: 499,
+    reviewsCount: 1,
+    totalScore: 1,
   },
   {
     id: '6',
@@ -172,10 +178,13 @@ export const products = [
     price: 7900,
   },
   {
+    brand: 'Tovar',
     id: '8',
     title: 'Товар',
     image: require('../images/mock/tshirt.png'),
     price: 10,
+    reviewsCount: 2,
+    totalScore: 4,
   },
 ];
