+++ mobile/src/shared/ui/Input/Input.styles.ts
@@ -1,0 +1,14 @@
+import {StyleSheet} from 'react-native';
+import {Colors} from 'shared/libs/helpers';
+
+export const InputStyles = StyleSheet.create({
+  input: {
+    backgroundColor: Colors.Gray300,
+    borderRadius: 12,
+    paddingVertical: 13,
+    paddingHorizontal: 16,
+    width: 275,
+    height: 48,
+    marginBottom: 20,
+  },
+});
