+++ mobile/src/shared/ui/Promo/Promo.styles.ts
@@ -1,0 +1,12 @@
+import {StyleSheet} from 'react-native';
+
+export const PromoStyles = StyleSheet.create({
+  container: {
+    width: '100%',
+    position: 'relative',
+  },
+  bannerImage: {
+    width: '100%',
+    height: '100%',
+  },
+});
