+++ mobile/src/widgets/Header/ui/index.tsx
@@ -5,6 +5,7 @@ import React from 'react';
 import {View, TouchableOpacity, Text, StyleSheet} from 'react-native';
 import {CartIcon} from 'shared/icons/Icons';
 import {Colors, IconStyles, TextStyles} from 'shared/libs/helpers';
+import {HeaderStyles as styles} from './Header.styles';
 
 type HeaderProps = {
   navigation: NavigationProp<RootStackParamsList>;
@@ -34,16 +35,3 @@ export const Header = ({navigation}: HeaderProps) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    height: 60,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: 15,
-    backgroundColor: Colors.White100,
-    borderBottomWidth: 2,
-    borderBottomColor: Colors.Gray100,
-  },
-});
