+++ mobile/src/widgets/Slider/ui/Slider.styles.ts
@@ -1,0 +1,17 @@
+import {StyleSheet} from 'react-native';
+
+export const SliderStyles = StyleSheet.create({
+  flatList: {
+    gap: 10,
+    paddingHorizontal: 10,
+  },
+  container: {
+    gap: 10,
+  },
+  header: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'baseline',
+    paddingHorizontal: 10,
+  },
+});
